Changed Calculation Type for MatrixMultiplier from double to float for compatibilty and performance

+++ Inc/ST-LIB_HIGH/Control/Blocks/MatrixMultiplier.hpp
@@ -4,10 +4,10 @@
 template<size_t U, size_t V, size_t W> class MatrixMultiplier;
 
 template<size_t U, size_t V, size_t W> requires (U > 1 && V > 1 && W > 1)
-class MatrixMultiplier<U,V,W> : public ControlBlock<double(*)[U][V], double(*)[U][W]>{
-	double(&A)[U][V], (&B)[V][W], (&result)[U][W];
+class MatrixMultiplier<U,V,W> : public ControlBlock<float(*)[U][V], float(*)[U][W]>{
+	float(&A)[U][V], (&B)[V][W], (&result)[U][W];
 public:
-	MatrixMultiplier(double(&A)[U][V], double(&B)[V][W], double(&result)[U][W]) :  A(A), B(B), result(result){
+	MatrixMultiplier(float(&A)[U][V], float(&B)[V][W], float(&result)[U][W]) :  A(A), B(B), result(result){
         this->input_value = &A;
         this->output_value = &result;
 	}
@@ -33,14 +33,14 @@ public:
 //CTAD
 #if __cpp_deduction_guides >= 201606
 template<size_t U, size_t V, size_t W> requires (U > 1 && V > 1 && W > 1)
-MatrixMultiplier(double(&)[U][V], double(&)[V][W], double(&)[U][W])->MatrixMultiplier<U,V,W>;
+MatrixMultiplier(float(&)[U][V], float(&)[V][W], float(&)[U][W])->MatrixMultiplier<U,V,W>;
 #endif
 
 template<size_t U, size_t V>
-class MatrixMultiplier<U,V,1> : public ControlBlock<double(*)[U][V], double(*)[U]>{
-    double(&A)[U][V], (&B)[V], (&result)[U];
+class MatrixMultiplier<U,V,1> : public ControlBlock<float(*)[U][V], float(*)[U]>{
+    float(&A)[U][V], (&B)[V], (&result)[U];
 public:
-    MatrixMultiplier(double(&A)[U][V], double(&B)[V], double(&result)[U]) :  A(A), B(B), result(result){
+    MatrixMultiplier(float(&A)[U][V], float(&B)[V], float(&result)[U]) :  A(A), B(B), result(result){
         this->input_value = &A;
         this->output_value = &result;
     }
@@ -63,14 +63,14 @@ public:
 //CTAD
 #if __cpp_deduction_guides >= 201606
 template<size_t U, size_t V> requires (U > 1 && V > 1)
-MatrixMultiplier(double(&)[U][V], double(&)[V], double(&)[U])->MatrixMultiplier<U,V,1>;
+MatrixMultiplier(float(&)[U][V], float(&)[V], float(&)[U])->MatrixMultiplier<U,V,1>;
 #endif
 
 template<size_t U, size_t V>
-class MatrixMultiplier<1,U,V> : public ControlBlock<double(*)[U], double(*)[V]>{
-    double(&A)[U], (&B)[U][V], (&result)[V];
+class MatrixMultiplier<1,U,V> : public ControlBlock<float(*)[U], float(*)[V]>{
+    float(&A)[U], (&B)[U][V], (&result)[V];
 public:
-    MatrixMultiplier(double(&A)[U], double(&B)[U][V], double(&result)[V]) :  A(A), B(B), result(result){
+    MatrixMultiplier(float(&A)[U], float(&B)[U][V], float(&result)[V]) :  A(A), B(B), result(result){
         this->input_value = &A;
         this->output_value = &result;
     }
@@ -93,14 +93,14 @@ public:
 //CTAD
 #if __cpp_deduction_guides >= 201606
 template<size_t U, size_t V> requires (U > 1 && V > 1)
-MatrixMultiplier(double(&)[U], double(&)[U][V], double(&)[V])->MatrixMultiplier<1,U,V>;
+MatrixMultiplier(float(&)[U], float(&)[U][V], float(&)[V])->MatrixMultiplier<1,U,V>;
 #endif
 
 template<size_t U>
-class MatrixMultiplier<1,U,1> : public ControlBlock<double(*)[U], double(*)>{
-    double(&A)[U], (&B)[U], (&result);
+class MatrixMultiplier<1,U,1> : public ControlBlock<float(*)[U], float(*)>{
+    float(&A)[U], (&B)[U], &result;
 public:
-    MatrixMultiplier(double(&A)[U], double(&B)[U], double(&result)) :  A(A), B(B), result(result){
+    MatrixMultiplier(float(&A)[U], float(&B)[U], float(&result)) :  A(A), B(B), result(result){
         this->input_value = &A;
         this->output_value = &result;
     }
@@ -119,5 +119,5 @@ public:
 //CTAD
 #if __cpp_deduction_guides >= 201606
 template<size_t U> requires (U > 1)
-MatrixMultiplier(double(&)[U], double(&)[U], double(&))->MatrixMultiplier<1,U,1>;
+MatrixMultiplier(float(&)[U], float(&)[U], float(&))->MatrixMultiplier<1,U,1>;
 #endif
